do not show nav links when not logged in
